Refactor terminal interface and command definition
The terminal layer is now split from the command processing backend - this
means, that parameters are parsed and validated before the actual command
routine is run; the same applies to output parameters - they are set to a
container and external logic converts them to desired representation.

Error processing is now also generalized - the command may set an error
return value to indicate a failure.

Furthermore, command flags has been added; for now, its main purpose is to
differentiate command group from command, but in future, one may specify
the command to be run from local terminal-only.

This is a first step towards remote terminal implementation; this commit
is related to issue #4